# build/contracts/dotnet/staff.v1.openapi.json
{
  "components": {
    "schemas": {
...
      "CursorEnvelopeOfApprovalQueueEntry": {
        "description": "The wire shape of a cursor-paged collection.",
        "properties": {
          "items": {
            "description": "The items in this page.",
            "items": {
              "$ref": "#/components/schemas/ApprovalQueueEntry"
            },
            "type": "array"
          },
          "nextCursor": {
            "description": "An opaque cursor to resume from, or \u0060null\u0060 when the collection is exhausted.",
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "CursorEnvelopeOfUnexecutedApproval": {
        "description": "The wire shape of a cursor-paged collection.",
        "properties": {
          "items": {
            "description": "The items in this page.",
            "items": {
              "$ref": "#/components/schemas/UnexecutedApproval"
            },
            "type": "array"
          },
          "nextCursor": {
            "description": "An opaque cursor to resume from, or \u0060null\u0060 when the collection is exhausted.",
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "ProblemContract": {
        "description": "The problem-details shape both deployables emit, so a client parses one error contract.",
        "properties": {
          "correlationId": {
            "description": "The correlation identifier for this request.",
            "type": "string"
          },
          "detail": {
            "description": "A client-safe explanation. Never an exception message.",
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "description": "The request path this problem occurred on.",
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "description": "The HTTP status code.",
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "title": {
            "description": "A short, client-safe summary.",
            "type": "string"
          },
          "type": {
            "description": "A ProblemTypes URI.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "correlationId"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The staff audience. Generated from the running service; not hand-maintained. Identity is derived at the gateway and no operation accepts a tenant, role or audience parameter.",
    "title": "Synthia Staff API",
    "version": "v1"
  },
  "openapi": "3.1.1",
  "paths": {
    "/api/staff/v1/views/approvals/queue": {
      "get": {
        "operationId": "StaffListApprovalQueue",
        "parameters": [
          {
            "description": "Page size. Clamped to the server maximum rather than refused.",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "An opaque resume position from a previous page\u0027s nextCursor. Clients MUST NOT construct or decode one.",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Sort field, prefixed with \u0027-\u0027 for descending. An unlisted field is a 400. Default: -createdAt.",
            "in": "query",
            "name": "sort",
            "schema": {
              "enum": [
                "-createdAt",
                "-expiresAt",
                "createdAt",
                "expiresAt"
              ],
              "type": "string"
            }
          },
          {
            "description": "Narrows the result to one organisation WITHIN the set the caller may already see. It does not establish authority and cannot widen the result: naming an organisation outside the caller\u0027s scope returns no rows.",
            "in": "query",
            "name": "tenantId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorEnvelopeOfApprovalQueueEntry"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Lists pending approvals with their fully disclosed commands. Cursor paged.",
        "tags": [
          "staff-views"
        ]
      }
    },
    "/api/staff/v1/views/approvals/unexecuted": {
      "get": {
        "operationId": "StaffListUnexecutedApprovals",
        "parameters": [
          {
            "description": "Page size. Clamped to the server maximum rather than refused.",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "An opaque resume position from a previous page\u0027s nextCursor. Clients MUST NOT construct or decode one.",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Sort field, prefixed with \u0027-\u0027 for descending. An unlisted field is a 400. Default: -decidedAt.",
            "in": "query",
            "name": "sort",
            "schema": {
              "enum": [
                "-decidedAt",
                "-expiresAt",
                "decidedAt",
                "expiresAt"
              ],
              "type": "string"
            }
          },
          {
            "description": "Narrows the result to one organisation WITHIN the set the caller may already see. It does not establish authority and cannot widen the result: naming an organisation outside the caller\u0027s scope returns no rows.",
            "in": "query",
            "name": "tenantId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorEnvelopeOfUnexecutedApproval"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Lists work approved but never executed. Cursor paged.",
        "tags": [
          "staff-views"
        ]
      }
    },
    "/api/staff/v1/views/audit": {
      "get": {
        "operationId": "StaffSearchAudit",
        "parameters": [
          {
            "description": "Page size. Clamped to the server maximum rather than refused.",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "An opaque resume position from a previous page\u0027s nextCursor. Clients MUST NOT construct or decode one.",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Sort field, prefixed with \u0027-\u0027 for descending. An unlisted field is a 400. Default: -occurredAt.",
            "in": "query",
            "name": "sort",
            "schema": {
              "enum": [
                "-occurredAt",
                "occurredAt"
              ],
              "type": "string"
            }
          },
          {
            "description": "Filters on eventKind. An unlisted filter field is a 400.",
            "in": "query",
            "name": "eventKind",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filters on occurredFrom. An unlisted filter field is a 400.",
            "in": "query",
            "name": "occurredFrom",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Filters on occurredTo. An unlisted filter field is a 400.",
            "in": "query",
            "name": "occurredTo",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Narrows the result to one organisation WITHIN the set the caller may already see. It does not establish authority and cannot widen the result: naming an organisation outside the caller\u0027s scope returns no rows.",
            "in": "query",
            "name": "tenantId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Filters on workItemId. An unlisted filter field is a 400.",
            "in": "query",
            "name": "workItemId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorEnvelopeOfAuditEventView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Searches audit records. Cursor paged.",
        "tags": [
          "staff-views"
        ]
      }
    },
    "/api/staff/v1/views/dashboard/platform": {
      "get": {
        "operationId": "StaffGetPlatformDashboard",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardRollup"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Reads the platform rollup. Aggregate only; no per-organisation breakdown.",
        "tags": [
          "staff-views"
        ]
      }
    },
...
    "/api/staff/v1/views/sessions/{sessionId}": {
      "get": {
        "operationId": "StaffGetSession",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffSessionDetail"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Reads one session and its step trail.",
        "tags": [
          "staff-views"
        ]
      }
    },
    "/api/staff/v1/views/tenants": {
      "get": {
        "operationId": "StaffListTenants",
        "parameters": [
          {
            "description": "Page size. Clamped to the server maximum rather than refused.",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "An opaque resume position from a previous page\u0027s nextCursor. Clients MUST NOT construct or decode one.",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Sort field, prefixed with \u0027-\u0027 for descending. An unlisted field is a 400. Default: displayName.",
            "in": "query",
            "name": "sort",
            "schema": {
              "enum": [
                "-displayName",
                "-status",
                "displayName",
                "status"
              ],
              "type": "string"
            }
          },
          {
            "description": "Filters on status. An unlisted filter field is a 400.",
            "in": "query",
            "name": "status",
            "schema": {
              "enum": [
                "active",
                "offboarded",
                "suspended",
                "unknown"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorEnvelopeOfTenantView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemContract"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Lists organisations. Cursor paged.",
        "tags": [
          "staff-views"
        ]
      }
    }
  },
...
}